Invert Filters
When games are returned from the API, all mechanisms are filtered out,
showing no games.  When selecting a filter, it will show all games that
contain that mechanism.  These filters are additive, meaning that as
long as the game has one of these mechanics, it will show in the game
list.